requirements.txt
basemap==1.2.0
cycler==0.10.0
dateparser==0.7.1
georinex==1.10.1
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
pyparsing==2.4.0
pyproj==2.2.0
pyshp==2.1.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
regex==2019.6.5
scipy==1.3.0
six==1.12.0
tec-embrace==0.2
tzlocal==1.5.1
xarray==0.12.1
